+++ src/rest/IntelligenceBase.ts
@@ -10,30 +10,24 @@
  */
 
 import Domain from "../base/Domain";
-import V1 from "./numbers/V1";
-import V2 from "./numbers/V2";
+import V2 from "./intelligence/V2";
 
-class NumbersBase extends Domain {
-  _v1?: V1;
+class IntelligenceBase extends Domain {
   _v2?: V2;
 
   /**
-   * Initialize numbers domain
+   * Initialize intelligence domain
    *
    * @param twilio - The twilio client
    */
   constructor(twilio: any) {
-    super(twilio, "https://numbers.twilio.com");
+    super(twilio, "https://intelligence.twilio.com");
   }
 
-  get v1(): V1 {
-    this._v1 = this._v1 || new V1(this);
-    return this._v1;
-  }
   get v2(): V2 {
     this._v2 = this._v2 || new V2(this);
     return this._v2;
   }
 }
 
-export = NumbersBase;
+export = IntelligenceBase;