+++ src/rest/supersim/v1/fleet.ts
@@ -331,12 +331,12 @@ export class FleetInstance {
    */
   dataEnabled: boolean;
   /**
-   * The total data usage (download and upload combined) in Megabytes that each Super SIM assigned to the Fleet can consume during a billing period (normally one month). Value must be between 1MB (1) and 2TB (2,000,000). Defaults to 1GB (1,000).
+   * The total data usage (download and upload combined) in Megabytes that each Super SIM assigned to the Fleet can consume during a billing period (normally one month). Value must be between 1MB (1) and 2TB (2,000,000). Defaults to 250MB.
    */
   dataLimit: number;
   dataMetering: FleetDataMetering;
   /**
-   * Defines whether SIMs in the Fleet are capable of sending and receiving machine-to-machine SMS via Commands. Defaults to `true`.
+   * Defines whether SIMs in the Fleet are capable of sending and receiving machine-to-machine SMS via Commands. Defaults to `false`.
    */
   smsCommandsEnabled: boolean;
   /**